fix: Modify format

+++ battery/interfaces/hdi_service/src/power_supply_provider.cpp
@@ -86,7 +86,7 @@ inline void PowerSupplyProvider::Trim(char* str)
         return;
     }
 
-    int32_t strc = strcspn(str, "\n");
+    size_t strc = strcspn(str, "\n");
     if (strc < 0 || strc >= strlen(str)) {
         return;
     }
